+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [20.14.5] - 2025-01-09
+
+### Fixed
+- Fixed the player can't play the same attack animation, until it fully ends in the server side
+
+### For Devs
+- ComboCounterHandler
+    - A more modularized and parameter sensitive way to handle combo counter of `BasicAttack`
+    - CapabilityItem.shouldCancelCombo is now deprecated, as more parameter sensitive version introduced
+
 ## [20.14.4] - 2025-12-17
 
 ### Changed